+++ server/src/emails/album-invite.email.tsx
@@ -2,7 +2,7 @@ import { Img, Link, Section, Text } from '@react-email/components';
 import * as React from 'react';
 import { ImmichButton } from 'src/emails/components/button.component';
 import ImmichLayout from 'src/emails/components/immich.layout';
-import { AlbumInviteEmailProps } from 'src/interfaces/notification.interface';
+import { AlbumInviteEmailProps } from 'src/repositories/notification.repository';
 import { replaceTemplateTags } from 'src/utils/replace-template-tags';
 
 export const AlbumInviteEmail = ({
