+++ server/src/emails/test.email.tsx
@@ -1,7 +1,7 @@
 import { Link, Row, Text } from '@react-email/components';
 import * as React from 'react';
 import ImmichLayout from 'src/emails/components/immich.layout';
-import { TestEmailProps } from 'src/interfaces/notification.interface';
+import { TestEmailProps } from 'src/repositories/notification.repository';
 
 export const TestEmail = ({ baseUrl, displayName }: TestEmailProps) => (
   <ImmichLayout preview="This is a test email from Immich.">
