+++ server/src/emails/welcome.email.tsx
@@ -2,7 +2,7 @@ import { Link, Section, Text } from '@react-email/components';
 import * as React from 'react';
 import { ImmichButton } from 'src/emails/components/button.component';
 import ImmichLayout from 'src/emails/components/immich.layout';
-import { WelcomeEmailProps } from 'src/interfaces/notification.interface';
+import { WelcomeEmailProps } from 'src/repositories/notification.repository';
 import { replaceTemplateTags } from 'src/utils/replace-template-tags';
 
 export const WelcomeEmail = ({ baseUrl, displayName, username, password, customTemplate }: WelcomeEmailProps) => {
