+++ server/src/repositories/notification.repository.ts
@@ -6,18 +6,104 @@ import { AlbumInviteEmail } from 'src/emails/album-invite.email';
 import { AlbumUpdateEmail } from 'src/emails/album-update.email';
 import { TestEmail } from 'src/emails/test.email';
 import { WelcomeEmail } from 'src/emails/welcome.email';
-import {
-  EmailRenderRequest,
-  EmailTemplate,
-  INotificationRepository,
-  SendEmailOptions,
-  SendEmailResponse,
-  SmtpOptions,
-} from 'src/interfaces/notification.interface';
 import { LoggingRepository } from 'src/repositories/logging.repository';
 
+export type EmailImageAttachment = {
+  filename: string;
+  path: string;
+  cid: string;
+};
+
+export type SendEmailOptions = {
+  from: string;
+  to: string;
+  replyTo?: string;
+  subject: string;
+  html: string;
+  text: string;
+  imageAttachments?: EmailImageAttachment[];
+  smtp: SmtpOptions;
+};
+
+export type SmtpOptions = {
+  host: string;
+  port?: number;
+  username?: string;
+  password?: string;
+  ignoreCert?: boolean;
+};
+
+export enum EmailTemplate {
+  TEST_EMAIL = 'test',
+
+  // AUTH
+  WELCOME = 'welcome',
+  RESET_PASSWORD = 'reset-password',
+
+  // ALBUM
+  ALBUM_INVITE = 'album-invite',
+  ALBUM_UPDATE = 'album-update',
+}
+
+interface BaseEmailProps {
+  baseUrl: string;
+  customTemplate?: string;
+}
+
+export interface TestEmailProps extends BaseEmailProps {
+  displayName: string;
+}
+
+export interface WelcomeEmailProps extends BaseEmailProps {
+  displayName: string;
+  username: string;
+  password?: string;
+}
+
+export interface AlbumInviteEmailProps extends BaseEmailProps {
+  albumName: string;
+  albumId: string;
+  senderName: string;
+  recipientName: string;
+  cid?: string;
+}
+
+export interface AlbumUpdateEmailProps extends BaseEmailProps {
+  albumName: string;
+  albumId: string;
+  recipientName: string;
+  cid?: string;
+}
+
+export type EmailRenderRequest =
+  | {
+      template: EmailTemplate.TEST_EMAIL;
+      data: TestEmailProps;
+      customTemplate: string;
+    }
+  | {
+      template: EmailTemplate.WELCOME;
+      data: WelcomeEmailProps;
+      customTemplate: string;
+    }
+  | {
+      template: EmailTemplate.ALBUM_INVITE;
+      data: AlbumInviteEmailProps;
+      customTemplate: string;
+    }
+  | {
+      template: EmailTemplate.ALBUM_UPDATE;
+      data: AlbumUpdateEmailProps;
+      customTemplate: string;
+    };
+
+export type SendEmailResponse = {
+  messageId: string;
+  response: any;
+};
+
 @Injectable()
-export class NotificationRepository implements INotificationRepository {
+export class NotificationRepository {
   constructor(private logger: LoggingRepository) {
     this.logger.setContext(NotificationRepository.name);
   }
