+++ server/src/types.ts
@@ -12,6 +12,7 @@ import { MapRepository } from 'src/repositories/map.repository';
 import { MediaRepository } from 'src/repositories/media.repository';
 import { MemoryRepository } from 'src/repositories/memory.repository';
 import { MetadataRepository } from 'src/repositories/metadata.repository';
+import { NotificationRepository } from 'src/repositories/notification.repository';
 import { OAuthRepository } from 'src/repositories/oauth.repository';
 import { ServerInfoRepository } from 'src/repositories/server-info.repository';
 import { MetricGroupRepository, TelemetryRepository } from 'src/repositories/telemetry.repository';
@@ -55,6 +56,7 @@ export type IMediaRepository = RepositoryInterface<MediaRepository>;
 export type IMemoryRepository = RepositoryInterface<MemoryRepository>;
 export type IMetadataRepository = RepositoryInterface<MetadataRepository>;
 export type IMetricGroupRepository = RepositoryInterface<MetricGroupRepository>;
+export type INotificationRepository = RepositoryInterface<NotificationRepository>;
 export type IOAuthRepository = RepositoryInterface<OAuthRepository>;
 export type IServerInfoRepository = RepositoryInterface<ServerInfoRepository>;
 export type ITelemetryRepository = RepositoryInterface<TelemetryRepository>;
