+++ server/test/repositories/notification.repository.mock.ts
@@ -1,4 +1,4 @@
-import { INotificationRepository } from 'src/interfaces/notification.interface';
+import { INotificationRepository } from 'src/types';
 import { Mocked } from 'vitest';
 
 export const newNotificationRepositoryMock = (): Mocked<INotificationRepository> => {
