+++ server/test/utils.ts
@@ -13,6 +13,7 @@ import { MapRepository } from 'src/repositories/map.repository';
 import { MediaRepository } from 'src/repositories/media.repository';
 import { MemoryRepository } from 'src/repositories/memory.repository';
 import { MetadataRepository } from 'src/repositories/metadata.repository';
+import { NotificationRepository } from 'src/repositories/notification.repository';
 import { OAuthRepository } from 'src/repositories/oauth.repository';
 import { ServerInfoRepository } from 'src/repositories/server-info.repository';
 import { TelemetryRepository } from 'src/repositories/telemetry.repository';
@@ -32,6 +33,7 @@ import {
   IMediaRepository,
   IMemoryRepository,
   IMetadataRepository,
+  INotificationRepository,
   IOAuthRepository,
   IServerInfoRepository,
   ITrashRepository,
@@ -157,7 +159,7 @@ export const newTestService = <T extends BaseService>(
     memoryMock as IMemoryRepository as MemoryRepository,
     metadataMock as IMetadataRepository as MetadataRepository,
     moveMock,
-    notificationMock,
+    notificationMock as INotificationRepository as NotificationRepository,
     oauthMock as IOAuthRepository as OAuthRepository,
     partnerMock,
     personMock,
